feat: Add mock provider
