+++ ImageFeed/Screens/ImagesList/Model/Picture.swift
@@ -12,16 +12,3 @@ struct Picture {
     let date: Date
     var isFavorite: Bool
 }
-
-// MARK: - MockData
-extension Picture {
-    static let pictures: [Picture] = {
-        (0...20).map { number in
-            Picture(
-                image: String(describing: number),
-                date: Date(),
-                isFavorite: number % 2 == 0
-            )
-        }
-    }()
-}
